When hitting a branch instruction with jumps remaining, the instruction pointer is set to point to the instruction after the label which was branched to.  This should help nested branches work to some degree.  Basically, what was working still works, but nested branches should have some functionality.
NOTE: The number of times a branch can be followed will NOT reset, even if a jump from another branch would jump to before the branch.  So y'know, be aware.

Modified SimulationForm.cs and how branch instructions are setup.
Added an AddInstruction method to more effectively add instructions to an expanded instruction list.
Added an AddBranch method to more effectively add branches to an expanded instruction list.